requirements.txt
Flask==0.10.1
Flask-Assets==0.9
Flask-Bootstrap==3.1.1.2
Flask-Login==0.2.10
Flask-Mail==0.9.0
Flask-OAuth==0.12
Flask-Principal==0.4.0
Flask-PyMongo==0.3.0
Flask-SQLAlchemy==1.0
Flask-Script==0.6.7
Flask-Security==1.7.1
Flask-Social==1.6.2
Flask-WTF==0.9.5
Jinja2==2.7.2
MarkupSafe==0.19
PyYAML==3.11
SQLAlchemy==0.9.4
WTForms==1.0.5
Werkzeug==0.9.4
amqp==1.4.4
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.16
blinker==1.3
celery==3.1.10
facebook-sdk==0.4.0
gunicorn==18.0
httplib2==0.8
itsdangerous==0.24
kombu==3.0.14
mock==1.0.1
oauth2==1.5.211
oauthlib==0.6.1
passlib==1.6.2
petl==0.23
psycopg2==2.5.2
pymongo==2.7
python-tumblpy==1.0.2
pytz==2014.2
requests==2.1.0
requests-oauthlib==0.4.0
twython==3.1.2
webassets==0.9
wsgiref==0.1.2
